+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## FIWARE Big Bang v0.39.0-next
 
+-   Fix maintenance script (#393)
+
 ## FIWARE Big Bang v0.39.0 - 09 June, 2024
 
 -   Update Mongodb to 6.0 for Orion (#389)